scripts/reddit/requirements.txt
praw
ecs-logging
python-dotenv